requirements.txt
discord.py
google
googlesearch-python
beautifulsoup4
requests
Flask
python-dotenv